add vscode; clean up shellHooks;

+++ shell.nix
@@ -1,10 +1,6 @@
 { pkgs ? (import ./nix/sources.nix).nixpkgs }:
 let
 
-  inherit (pkgs)
-    asdf-vm
-    ;
-
   systemSet = {
     inherit (pkgs)
       # coreutils-full git zsh oh-my-zsh
@@ -14,7 +10,7 @@ let
 
   devtoolsSet = {
     inherit (pkgs)
-      # vim
+      # neovim
       docker docker-compose# kubernetes minikube k9s
       jq dnsmasq sqlite-interactive# or just sqlite?
       nmap socat dig inetutils# ngrok
@@ -33,60 +29,63 @@ let
 
   inherit (pkgs) mkShell;
 
-  shellHook = ''
-    cd $NIV_ROOT
-    niv update nixpkgs
-    cd -
-  '';
+  # shellHook = ''
+  #   #!${pkgs.runtimeShell}
+  #   cd $NIV_ROOT
+  #   niv update nixpkgs
+  #   cd -
+  # '';
 
-  NIV_ROOT = builtins.toString ./.;
+  # NIV_ROOT = builtins.toString ./.;
 
 in
 rec {
 
   systemShell = mkShell {
     buildInputs = systemList;
-
-    inherit shellHook NIV_ROOT;
   };
 
   devtoolsShell = mkShell {
     buildInputs = devtoolsList;
-
-    inherit shellHook NIV_ROOT;
   };
 
   golangShell = mkShell {
     buildInputs = golangList;
-
-    inherit shellHook NIV_ROOT;
   };
 
   completeShell = mkShell {
-    buildInputs = systemList ++ devtoolsList ++ golangList;
+    buildInputs = systemList ++ devtoolsList;
     shellHook = ''
-      asdf plugin add nodejs https://github.com/asdf-vm/asdf-nodejs.git
-      asdf install nodejs latest
-      asdf global nodejs latest
+      #!${pkgs.runtimeShell}
+      if [[ ! -e $(which node) ]]; then
+        asdf plugin add nodejs https://github.com/asdf-vm/asdf-nodejs.git
+        asdf install nodejs latest
+        asdf global nodejs latest
+      fi
 
-      asdf plugin-add golang https://github.com/kennyp/asdf-golang.git
-      asdf install golang latest
-      asdf global golang latest
+      if [[ ! -e $(which go) ]]; then
+        asdf plugin-add golang https://github.com/kennyp/asdf-golang.git
+        asdf install golang latest
+        asdf global golang latest
+      fi
     '';
-
-    inherit NIV_ROOT;
   };
 
   testShell = mkShell {
-    buildInputs = [ asdf-vm ];
+    buildInputs = [ pkgs.asdf-vm ];
     shellHook = ''
-      asdf plugin add nodejs https://github.com/asdf-vm/asdf-nodejs.git
-      asdf install nodejs latest
-      asdf global nodejs latest
+      #!${pkgs.runtimeShell}
+      if [[ ! -e $(which node) ]]; then
+        asdf plugin add nodejs https://github.com/asdf-vm/asdf-nodejs.git
+        asdf install nodejs latest
+        asdf global nodejs latest
+      fi
 
-      asdf plugin-add golang https://github.com/kennyp/asdf-golang.git
-      asdf install golang latest
-      asdf global golang latest
+      if [[ ! -e $(which go) ]]; then
+        asdf plugin-add golang https://github.com/kennyp/asdf-golang.git
+        asdf install golang latest
+        asdf global golang latest
+      fi
     '';
   };
